+++ performance/benchmark.rb
@@ -5,12 +5,14 @@
 
 require "benchmark/ips"
 
-# Configure Rails Envinronment
+# Configure Rails Environment
 ENV["RAILS_ENV"] = "production"
 require File.expand_path("../test/sandbox/config/environment.rb", __dir__)
 
-require_relative "components/name_component.rb"
-require_relative "components/inline_component.rb"
+module Performance
+  require_relative "components/name_component.rb"
+  require_relative "components/inline_component.rb"
+end
 
 class BenchmarksController < ActionController::Base
 end
@@ -22,8 +24,8 @@ Benchmark.ips do |x|
   x.time = 10
   x.warmup = 2
 
-  x.report("component:") { controller_view.render(NameComponent.new(name: "Fox Mulder")) }
-  x.report("inline:") { controller_view.render(InlineComponent.new(name: "Fox Mulder")) }
+  x.report("component:") { controller_view.render(Performance::NameComponent.new(name: "Fox Mulder")) }
+  x.report("inline:") { controller_view.render(Performance::InlineComponent.new(name: "Fox Mulder")) }
   x.report("partial:") { controller_view.render("partial", name: "Fox Mulder") }
 
   x.compare!